# openapi.yaml
openapi: 3.0.3
info:
  title: Translation Management API
  version: 1.0.0
paths:
  /api/translations:
    post:
      summary: Create a new translation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationDTO'
      responses:
        '200':
          description: Translation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationDTO'
  /api/translations/{id}:
    put:
      summary: Update a translation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationDTO'
      responses:
        '200':
          description: Translation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationDTO'
    get:
      summary: Get a translation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Translation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationDTO'
  /api/translations/search:
    get:
      summary: Search translations by keyword
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of translations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslationDTO'
  /api/translations/locale/{locale}:
    get:
      summary: Get translations by locale
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of translations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslationDTO'
  /api/translations/export/{locale}:
    get:
      summary: Export translations as JSON
      parameters:
        - name: locale
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: JSON export of translations
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    TranslationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        locale:
          type: string
        content:
          type: string
        tags:
          type: string